Create geotrellis.vector.mesh package
