+++ raster/src/main/scala/geotrellis/raster/triangulation/DelaunayRasterizer.scala
@@ -1,9 +1,9 @@
 package geotrellis.raster.triangulation
 
 import com.vividsolutions.jts.geom.Coordinate
-
 import geotrellis.raster._
 import geotrellis.vector._
+import geotrellis.vector.mesh.{HalfEdgeTable, IndexedPointSet}
 import geotrellis.vector.triangulation._
 
 object DelaunayRasterizer {
