+++ vector/src/main/scala/geotrellis/vector/mesh/IndexedPointSet.scala
@@ -1,7 +1,6 @@
-package geotrellis.vector.triangulation
+package geotrellis.vector.mesh
 
 import com.vividsolutions.jts.geom.Coordinate
-
 import geotrellis.util.Constants.{DOUBLE_EPSILON => EPSILON}
 import geotrellis.vector.Point
